refactoring ComputerVisionModule

+++ Source/SmartCompanion/ConfigData.h
@@ -6,5 +6,5 @@ inline const std::string baseDir = "D:\\SmartCompanion\\SmartCompanion";
 
 inline const size_t SPEECH_BUFFER_SIZE = 4096;
 
-inline const int IMG_WIDTH = 640;
-inline const int IMG_HEIGHT = 359;
+inline const float IMG_WIDTH = 640.f;
+inline const float IMG_HEIGHT = 359.f;
